Updated Optimised code for factorial of a number
The optimised technique used is termed as Half Successive Multiplication Algorithm.  Here the number of multiplication has been reduced to half. This works well for mid range numbers. It is still a costly algorithm on low range numbers. I have even added the time module so that when you run the code, you can see the time taken by the program for the execution.